+++ next/src/app/lib/constants.ts
@@ -1,0 +1,32 @@
+export const APP_NAME = "XpenseTrackr";
+
+export const PAGES = {
+  HOME: {
+    href: "/",
+    label: "Home",
+  },
+  PRICING: {
+    href: "/pricing",
+    label: "Pricing",
+  },
+  BLOG: {
+    href: "/blog",
+    label: "Blog",
+  },
+  ABOUT: {
+    href: "/about",
+    label: "About",
+  },
+};
+
+// TODO: reconcile with PAGES in MyAppBar
+export const FOOTER: any = {
+  PRIVACY: {
+    href: "/privacy",
+    label: "Privacy",
+  },
+  TERMS: {
+    href: "/tos",
+    label: "Terms of Service",
+  },
+};
